AeCanvas: add "explicit border path" API for outskirts outside and inside.
This is an optimization, so a user that knows how to build the Cairo path not
only for the background but also por the border. This way, we avoid to perform
the "double-width clipped border" technique to draw borders with outside
and inside outskirts.
This was an almost internal refactoring, ie that shouldn't affect users, except of
setPathKey:hasJoins:hasCaps:block:, converted to 
setPathWith:nilOrBorderPathWith:hasJoins:hasCaps: