Rename several unbounded* to unowned*
It should not affect library users as as the unowned are "wrapped" by
another method that sends autoRelease to avoid memory leaks, and
that's what the users send.

unbounded* occurrences were only a mistake, I always tried to write *unowned,
because the senders is responsible of freeing the returned instance.